#14: Create test partitioner for JUnit5 Repositories on leader (#21)
### Summary

This PR creates test discovery and partitioning systems. Currently, the
tests are partitioned "evenly" onto different workers without regard for
execution time. "Evenly" is defined as each worker having roughly the
same number of tests to execute.

This PR also adds a new environment variable to the leader,
`SPEED_NUM_WORKERS`, which is the number of workers to spawn. This can
be made dynamic at a later day but is manually set by the user at leader
execution time.

### Known Issues
While not a bug, the code between the `repository` packages in the
leaders and workers is extremely similar. We should consider a refactor
at a later time.

### Assumptions Made 
- For the leader we assume that in the provided TEST_PATH directory to
look through, there is eventually some subdirectory `/test/java` or
`/java/test` with the `.class` files for testing specifically
- For the worker we assume that in the provided TEST_PATH directory to
look through, there is eventually some subdirectory `/test/java` or
`/java/test` with the `.class files `for testing specifically and also
that there is eventually some subdirectory `/main/java` or `/java/main`
with the source `.class` files
- We assume a JUnit class to be a class in the testing path where in the
class has at least one method that has the `@Test` annotation

Issue: #14

---------

Co-authored-by: Owen X. Ralbovsky <oxr2@cornell.edu>